Hotfix/#1
- 기획 스탭 말풍선 버그 수정
- 말풍선 코인 수령 버그 수정
- 진행 상항 표시 바 버그 수정

+++ Stage_OnAir/Assets/02_Ingame/Scripts/Manager/GameManager.cs
@@ -345,7 +345,7 @@ public class GameManager : Singleton<GameManager>
             }
             if (data.Keys.Contains("BestProfit"))
             {
-                Best_Profit = int.Parse(data["BestProfit"]["N"].ToString());
+                Best_Profit = long.Parse(data["BestProfit"]["N"].ToString());
             }
 
             Debug.Log("기존 데이터 불러오기");
